requirements.txt
ads
annotated-types
anthropic
anyio
argon2-cffi
argon2-cffi-bindings
arrow
astropy
astroquery
asttokens
async-lru
attrs
babel
backcall
backports.tarfile
beautifulsoup4
bettermoments
bleach
bs4
casa-formats-io
casaconfig
casatools
certifi
cffi
charset-normalizer
click
cloudpickle
cmasher
colorspacious
comm
contourpy
cryptography
cycler
dask
debugpy
decorator
defusedxml
distro
docstring_parser
dotenv
e13tools
emcee
exceptiongroup
executing
fastjsonschema
fonttools
fqdn
fsspec
gofish
google-ai-generativelanguage
google-api-core
google-auth
google-genai
google-generativeai
googleapis-common-protos
grpcio
grpcio-status
h11
html5lib
httpcore
httpx
idna
importlib_metadata
importlib_resources
iniconfig
ipykernel
ipython
ipywidgets
isoduration
jaraco.classes
jaraco.context
jaraco.functools
jedi
jeepney
Jinja2
jiter
joblib
json5
jsonpointer
jsonschema
jsonschema-specifications
jupyter
jupyter-console
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab_pygments
jupyterlab_server
jupyterlab_widgets
keyring
kiwisolver
locket
MarkupSafe
matplotlib
matplotlib-inline
mistune
mock
more-itertools
nbclient
nbconvert
nbformat
nest-asyncio
notebook
notebook_shim
numpy
openai
overrides
packaging
pandas
pandocfilters
parso
partd
pexpect
pickleshare
pillow
pkgutil_resolve_name
platformdirs
pluggy
prometheus_client
prompt_toolkit
proto-plus
protobuf
psutil
ptyprocess
pure_eval
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic_core
pyerfa
Pygments
PyMuPDF
pyparsing
pytest
python-dateutil
python-dotenv
python-json-logger
pytz
pyvo
PyYAML
pyzmq
radio-beam
referencing
requests
rfc3339-validator
rfc3986-validator
rpds-py
rsa
scikit-learn
scipy
seaborn
SecretStorage
Send2Trash
six
sniffio
soupsieve
spectral-cube
stack-data
terminado
threadpoolctl
tinycss2
tomli
toolz
tornado
tqdm
traitlets
types-python-dateutil
typing_extensions
tzdata
uri-template
urllib3
wcwidth
webcolors
webencodings
websocket-client
Werkzeug
widgetsnbextension
zeus-mcmc
zipp
